docs: separate tabs
