+++ demo/src/components/playground/tab/AnnotationTab.tsx
@@ -1,15 +1,9 @@
 import React from "react";
-import * as THREE from "three";
 import { Context } from "../context";
 
 import RemoveIcon from "@site/static/icon/remove.svg";
-import { FaceAnnotation } from "../../../../../src";
 
 class AnnotationTab extends React.Component {
-  public componentDidMount() {
-    this._listenAnnotationAdd();
-  }
-
   public render() {
     const { state } = this.context;
     const view3D = state.view3D;
@@ -73,59 +67,6 @@ class AnnotationTab extends React.Component {
 
     URL.revokeObjectURL(dataURL);
   };
-
-  private _listenAnnotationAdd() {
-    const { state } = this.context;
-    const view3D = state.view3D!;
-    const canvas = view3D.renderer.canvas;
-    const raycaster = new THREE.Raycaster();
-
-    canvas.addEventListener("dblclick", evt => {
-      const model = view3D.model;
-      if (!model) return;
-
-      const pointer = new THREE.Vector2();
-      const devicePixelRatio = window.devicePixelRatio;
-      pointer.x = (evt.offsetX / canvas.width) * 2 * devicePixelRatio - 1;
-      pointer.y = -(evt.offsetY / canvas.height) * 2 * devicePixelRatio + 1;
-
-      raycaster.setFromCamera(pointer, view3D.camera.threeCamera);
-
-      const intersects = raycaster.intersectObjects([model.scene]);
-
-      if (!intersects.length) return;
-
-      const currentPose = view3D.camera.currentPose;
-      const el = document.createElement("div");
-      el.classList.add("view3d-annotation");
-      el.classList.add("default");
-
-      const tooltip = document.createElement("div");
-      tooltip.classList.add("view3d-annotation-tooltip");
-      tooltip.classList.add("default");
-      el.appendChild(tooltip);
-
-      view3D.annotation.wrapper.appendChild(el);
-
-      const intersect = intersects[0];
-      const size = view3D.renderer.size;
-      const aspect = Math.max(size.height / size.width, 1);
-
-      const newAnnotation = new FaceAnnotation(view3D, {
-        element: el,
-        baseFov: view3D.camera.baseFov,
-        baseDistance: view3D.camera.baseDistance,
-        aspect,
-        focus: [currentPose.yaw, currentPose.pitch, currentPose.zoom],
-        meshIndex: model.meshes.findIndex(mesh => mesh === intersects[0].object),
-        faceIndex: intersect.faceIndex
-      });
-      view3D.annotation.add(newAnnotation);
-      view3D.renderer.renderSingleFrame();
-
-      this.setState({});
-    });
-  }
 }
 
 AnnotationTab.contextType = Context;
