+++ demo/src/components/playground/tab/tabs.module.css
@@ -1,6 +1,7 @@
 .control {
   width: 358px;
   height: 100%;
+  flex-shrink: 0;
   display: flex;
   flex-direction: column;
   font-weight: 500;
